refactor: put all logic inside the store
